Added KDC As Dependency, not peerDependency
Apparently peerDependencies will count a global install, but you cannot
resolve a global install. This means that you have no real way (i
believe) of resolving the location of a globally installed kdc module.

So, by declaring it as a simple dependency we remove any issue with the
global junk.

Not quite as neat as a peerDependency, but it works.